Use the sharded marker eval in parsing
